Livewell-backend/requirements.txt
fastapi
google-auth
requests
pyjwt
uvicorn
supabase